Added some file structure by  moving Login Screen to its own package
