--- app/src/main/java/lk/ms/deadlines/screen/logingScreen.kt
@@ -1,16 +1,0 @@
-package lk.ms.deadlines.screen
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-
-class LoggingScreen : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-
-        }
-    }
-}
